requirements.txt
torch
torchvision
pandas
numpy
tqdm
pillow
scikit-image
opencv-python
# NAFNet Dependencies
lmdb
scipy
timm
fvcore
pyyaml